#0: transition from functional to class done to enable trace_2cq
