+++ models/experimental/functional_yolov8x/demo/demo.py
@@ -8,13 +8,14 @@ import sys
 import ttnn
 import torch
 import pytest
+from pathlib import Path
 import torch.nn as nn
 from loguru import logger
 from datetime import datetime
 from functools import partial
-from models.utility_functions import disable_persistent_kernel_cache
+from models.utility_functions import disable_persistent_kernel_cache, skip_for_grayskull
 from models.experimental.functional_yolov8x.reference import yolov8x_utils
-from models.experimental.functional_yolov8x.tt.class_ttnn_yolov8x import YOLOv8xModel
+from models.experimental.functional_yolov8x.tt.ttnn_yolov8x import YOLOv8xModel
 from models.experimental.functional_yolov8x.tt.ttnn_yolov8x_utils import custom_preprocessor
 from models.experimental.functional_yolov8x.demo.demo_utils import LoadImages, preprocess, postprocess
 
@@ -41,18 +42,45 @@ class Ensemble(nn.ModuleList):
         return y, None
 
 
+def attempt_download(file, repo="ultralytics/assets"):
+    tests = Path(__file__).parent.parent / "reference"
+    file_path = tests / Path(str(file).strip().replace("'", "").lower())
+    if not file_path.exists():
+        name = "yolov8x.pt"
+        msg = f"{file_path} missing, try downloading from https://github.com/{repo}/releases/"
+        try:
+            url = f"https://github.com/{repo}/releases/download/v8.3.0/{name}"
+            print(f"Downloading {url} to {file_path}...")
+            torch.hub.download_url_to_file(url, file_path)
+
+            assert file_path.exists() and file_path.stat().st_size > 1e6, f"Download failed for {name}"
+        except Exception as e:
+            print(f"Error downloading from GitHub: {e}. Trying secondary source...")
+
+            url = f"https://storage.googleapis.com/{repo}/ckpt/{name}"
+            print(f"Downloading {url} to {file_path}...")
+            os.system(f"curl -L {url} -o {file_path}")
+
+            if not file_path.exists() or file_path.stat().st_size < 1e6:
+                file_path.unlink(missing_ok=True)
+                print(f"ERROR: Download failure for {msg}")
+            else:
+                print(f"Download succeeded from secondary source!")
+    return file_path
+
+
 def attempt_load(weights, map_location=None):
     model = Ensemble()
     for w in weights if isinstance(weights, list) else [weights]:
-        w = "models/experimental/functional_yolov8x/demo/yolov8x.pt"
-        ckpt = torch.load(w, map_location=map_location)
+        weight_path = attempt_download(w)
+        print("Loading weights from:", weight_path)
+        ckpt = torch.load(weight_path, map_location=map_location)
         model.append(ckpt["ema" if ckpt.get("ema") else "model"].float().eval())
     for m in model.modules():
-        if type(m) in [nn.Hardswish, nn.LeakyReLU, nn.ReLU, nn.ReLU6, nn.SiLU]:
+        if isinstance(m, (nn.Hardswish, nn.LeakyReLU, nn.ReLU, nn.ReLU6, nn.SiLU)):
             m.inplace = True
-        elif type(m) is nn.Upsample:
+        elif isinstance(m, nn.Upsample):
             m.recompute_scale_factor = None
-
     if len(model) == 1:
         return model[-1]
     else:
@@ -95,19 +123,13 @@ def save_yolo_predictions_by_model(result, save_dir, image_path, model_name):
     print(f"Predictions saved to {output_path}")
 
 
+@skip_for_grayskull()
 @pytest.mark.parametrize("device_params", [{"l1_small_size": 32768}], indirect=True)
 @pytest.mark.parametrize(
     "source, model_type",
     [
-        # ("models/experimental/functional_yolov8x/demo/images/bus.jpg", "torch_model"),
+        ("models/experimental/functional_yolov8x/demo/images/bus.jpg", "torch_model"),
         ("models/experimental/functional_yolov8x/demo/images/bus.jpg", "tt_model"),
-        #     ("models/experimental/functional_yolov8x/demo/images/test1.jpg", "torch_model"),
-        #     ("models/experimental/functional_yolov8x/demo/images/test1.jpg", "tt_model"),
-        #     ("models/experimental/functional_yolov8x/demo/images/test2.jpg", "torch_model"),
-        #     ("models/experimental/functional_yolov8x/demo/images/test2.jpg", "tt_model"),
-        #     ("models/experimental/functional_yolov8x/demo/images/test3.jpg", "torch_model"),
-        #     ("models/experimental/functional_yolov8x/demo/images/test3.jpg", "tt_model"),
-        #
     ],
 )
 @pytest.mark.parametrize("res", [(640, 640)])
@@ -119,7 +141,7 @@ def test_demo(device, source, model_type, res):
         logger.info("Inferencing using Torch Model")
     else:
         state_dict = attempt_load("yolov8x.pt", map_location="cpu").state_dict()
-        parameters = custom_preprocessor(device, state_dict, inp_h=640, inp_w=640)
+        parameters = custom_preprocessor(device, state_dict, inp_h=res[0], inp_w=res[1])
         model = YOLOv8xModel(device=device, parameters=parameters)
         logger.info("Inferencing using ttnn Model")
 
@@ -219,7 +241,6 @@ def test_demo(device, source, model_type, res):
         im = preprocess(im0s, res=res)
 
         ttnn_im = im.permute((0, 2, 3, 1))
-        # ttnn_im= im.reshape(1, 1, ttnn_im.shape[0]*ttnn_im.shape[1]*ttnn_im.shape[2], ttnn_im.shape[3])
         ttnn_im = ttnn.from_torch(ttnn_im, dtype=ttnn.bfloat16)
 
         if model_type == "torch_model":
