+++ models/experimental/functional_yolov8x/tests/yolov8x_test_infra.py
@@ -1,15 +1,16 @@
 # SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
 
 # SPDX-License-Identifier: Apache-2.0
-
+import os
 import sys
 import ttnn
 import torch
 import torch.nn as nn
 from loguru import logger
 from functools import partial
+from pathlib import Path
 from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.experimental.functional_yolov8x.tt.class_ttnn_yolov8x import YOLOv8xModel
+from models.experimental.functional_yolov8x.tt.ttnn_yolov8x import YOLOv8xModel
 from models.experimental.functional_yolov8x.reference import yolov8x_utils
 from models.experimental.functional_yolov8x.tt.ttnn_yolov8x_utils import custom_preprocessor
 
@@ -41,18 +42,46 @@ class Ensemble(nn.ModuleList):
         return y, None
 
 
+def attempt_download(file, repo="ultralytics/assets"):
+    tests = Path(__file__).parent.parent / "reference"
+    file_path = tests / Path(str(file).strip().replace("'", "").lower())
+
+    if not file_path.exists():
+        name = "yolov8x.pt"
+        msg = f"{file_path} missing, try downloading from https://github.com/{repo}/releases/"
+        try:
+            url = f"https://github.com/{repo}/releases/download/v8.3.0/{name}"
+            print(f"Downloading {url} to {file_path}...")
+            torch.hub.download_url_to_file(url, file_path)
+
+            assert file_path.exists() and file_path.stat().st_size > 1e6, f"Download failed for {name}"
+        except Exception as e:
+            print(f"Error downloading from GitHub: {e}. Trying secondary source...")
+
+            url = f"https://storage.googleapis.com/{repo}/ckpt/{name}"
+            print(f"Downloading {url} to {file_path}...")
+            os.system(f"curl -L {url} -o {file_path}")
+
+            if not file_path.exists() or file_path.stat().st_size < 1e6:
+                file_path.unlink(missing_ok=True)
+                print(f"ERROR: Download failure for {msg}")
+            else:
+                print(f"Download succeeded from secondary source!")
+    return file_path
+
+
 def attempt_load(weights, map_location=None):
     model = Ensemble()
     for w in weights if isinstance(weights, list) else [weights]:
-        w = "models/experimental/functional_yolov8x/demo/yolov8x.pt"
-        ckpt = torch.load(w, map_location=map_location)
+        weight_path = attempt_download(w)
+        print("Loading weights from:", weight_path)
+        ckpt = torch.load(weight_path, map_location=map_location)
         model.append(ckpt["ema" if ckpt.get("ema") else "model"].float().eval())
     for m in model.modules():
-        if type(m) in [nn.Hardswish, nn.LeakyReLU, nn.ReLU, nn.ReLU6, nn.SiLU]:
+        if isinstance(m, (nn.Hardswish, nn.LeakyReLU, nn.ReLU, nn.ReLU6, nn.SiLU)):
             m.inplace = True
-        elif type(m) is nn.Upsample:
+        elif isinstance(m, nn.Upsample):
             m.recompute_scale_factor = None
-
     if len(model) == 1:
         return model[-1]
     else:
