+++ models/experimental/functional_yolov8x/tt/ttnn_yolov8x.py
@@ -6,536 +6,697 @@ import ttnn
 import math
 
 from models.experimental.functional_yolov8x.tt.ttnn_yolov8x_utils import (
-    autopad,
     ttnn_decode_bboxes,
 )
 
 
-def conv(
-    device,
-    x,
-    parameters,
-    path,
-    in_h,
-    in_w,
-    k=1,
-    s=1,
-    p=None,
-    g=1,
-    d=1,
-    act_block_h=False,
-    block_shard=None,
-    bfloat8=True,
-    change_shard=False,
-    deallocate_activation=False,
-    output_layout=ttnn.TILE_LAYOUT,
-    is_fused=True,
-    is_dfl=False,
-    is_detect_cv2=False,
-    width_shard=False,
-):
-    p = autopad(k, p, d)
-
-    conv_config = ttnn.Conv2dConfig(
-        dtype=ttnn.bfloat16,
-        weights_dtype=ttnn.bfloat16,
-        activation="",
-        shard_layout=ttnn.TensorMemoryLayout.HEIGHT_SHARDED,
-        input_channels_alignment=16 if x.shape[3] < 16 else 32,
+class Conv:
+    def __init__(
+        self,
+        device,
+        parameters,
+        path,
+        input_params,
+        groups=1,
+        dilation=1,
+        act_block_h=False,
+        block_shard=None,
+        bfloat8=True,
+        change_shard=False,
         deallocate_activation=False,
+        output_layout=ttnn.TILE_LAYOUT,
+        is_fused=True,
+        is_dfl=False,
+        is_detect_cv2=False,
+        width_shard=False,
+        act_blocks=32,
         enable_act_double_buffer=False,
         enable_split_reader=False,
-        enable_subblock_padding=False,
-        output_layout=output_layout,
-    )
+        reshard_if_not_optimal=False,
+        cache={},
+        batch_size=1,
+    ):
+        self.device = device
+        self.parameters = parameters
+        self.path = path
+        self.input_params = input_params
+        self.groups = groups
+        self.dilation = dilation
+        self.act_block_h = act_block_h
+        self.block_shard = block_shard
+        self.bfloat8 = bfloat8
+        self.change_shard = change_shard
+        self.deallocate_activation = deallocate_activation
+        self.output_layout = output_layout
+        self.is_fused = is_fused
+        self.is_dfl = is_dfl
+        self.is_detect_cv2 = is_detect_cv2
+        self.width_shard = width_shard
+        self.act_blocks = act_blocks
+        self.enable_act_double_buffer = enable_act_double_buffer
+        self.enable_split_reader = enable_split_reader
+        self.reshard_if_not_optimal = reshard_if_not_optimal
+        self.cache = cache
+        self.batch_size = batch_size
 
-    if deallocate_activation:
-        conv_config.deallocate_activation = deallocate_activation
+        self.conv_config = self._initialize_conv_config()
+        self.compute_config = self._initialize_compute_config()
+        self.weights, self.bias = self.parameters[path]
 
-    if change_shard:
-        conv_config.shard_layout = None
+    def _initialize_conv_config(self):
+        conv_config = ttnn.Conv2dConfig(
+            dtype=ttnn.bfloat16,
+            weights_dtype=ttnn.bfloat16,
+            activation="",
+            shard_layout=ttnn.TensorMemoryLayout.HEIGHT_SHARDED,
+            input_channels_alignment=16 if self.input_params[4] < 16 else 32,
+            act_block_w_div=1,
+            transpose_shards=False,
+            deallocate_activation=False,
+            enable_act_double_buffer=self.enable_act_double_buffer,
+            enable_split_reader=self.enable_split_reader,
+            enable_subblock_padding=False,
+            output_layout=self.output_layout,
+            reallocate_halo_output=False,
+            reshard_if_not_optimal=self.reshard_if_not_optimal,
+        )
 
-    if act_block_h:
-        conv_config.act_block_h_override = 32
+        if self.deallocate_activation:
+            conv_config.deallocate_activation = self.deallocate_activation
 
-    if block_shard:
-        conv_config.shard_layout = ttnn.TensorMemoryLayout.BLOCK_SHARDED
+        if self.change_shard:
+            conv_config.shard_layout = None
 
-    if width_shard:
-        conv_config.shard_layout = ttnn.TensorMemoryLayout.WIDTH_SHARDED
+        if self.act_block_h:
+            conv_config.act_block_h_override = self.act_blocks
 
-    compute_config = ttnn.init_device_compute_kernel_config(
-        device.arch(),
-        math_fidelity=ttnn.MathFidelity.LoFi,
-        fp32_dest_acc_en=False,
-        packer_l1_acc=False,
-    )
+        if self.block_shard:
+            conv_config.shard_layout = ttnn.TensorMemoryLayout.BLOCK_SHARDED
 
-    if is_fused:
-        fused_weight, fused_bias = parameters[path]
-    else:
-        conv_weight, conv_bias = parameters[path]
+        if self.width_shard:
+            conv_config.shard_layout = ttnn.TensorMemoryLayout.WIDTH_SHARDED
 
-    if bfloat8:
-        conv_config.weights_dtype = ttnn.bfloat8_b
+        if self.bfloat8:
+            conv_config.weights_dtype = ttnn.bfloat8_b
 
-    [x, [out_height, out_width]] = ttnn.conv2d(
-        input_tensor=x,
-        weight_tensor=fused_weight if is_fused else conv_weight,
-        in_channels=fused_weight.shape[1] if is_fused else conv_weight.shape[1],
-        out_channels=fused_weight.shape[0] if is_fused else conv_weight.shape[0],
-        device=device,
-        bias_tensor=fused_bias if is_fused else conv_bias,
-        kernel_size=(k, k),
-        stride=(s, s),
-        padding=(p, p),
-        dilation=(d, d),
-        batch_size=x.shape[0],
-        input_height=in_h,
-        input_width=in_w,
-        conv_config=conv_config,
-        compute_config=compute_config,
-        conv_op_cache={},
-        debug=False,
-        groups=g,
-        memory_config=None,
-        return_weights_and_bias=False,
-        return_output_dim=True,
-    )
+        return conv_config
 
-    if is_dfl:
-        return ttnn.reshape(x, (x.shape[0], 4, -1))
+    def _initialize_compute_config(self):
+        return ttnn.init_device_compute_kernel_config(
+            self.device.arch(),
+            math_fidelity=ttnn.MathFidelity.LoFi,
+            math_approx_mode=False,
+            fp32_dest_acc_en=False,
+            packer_l1_acc=False,
+        )
 
-    if is_detect_cv2:
-        x = ttnn.sharded_to_interleaved(x, ttnn.L1_MEMORY_CONFIG)
-        return (x, out_height, out_width)
+    def __call__(self, x):
+        if x.shape[1] != 1:
+            input_height = x.shape[1]
+            input_width = x.shape[2]
+        else:
+            input_height = int(math.sqrt(x.shape[2]) // self.batch_size)
+            input_width = int(math.sqrt(x.shape[2]) // self.batch_size)
 
-    x = ttnn.silu(x)
+        [x, [out_height, out_width], [self.weights, self.bias]] = ttnn.conv2d(
+            input_tensor=x,
+            weight_tensor=self.weights,
+            in_channels=self.input_params[4],
+            out_channels=self.input_params[3],
+            device=self.device,
+            bias_tensor=self.bias,
+            kernel_size=(self.input_params[0], self.input_params[0]),
+            stride=(self.input_params[1], self.input_params[1]),
+            padding=(self.input_params[2], self.input_params[2]),
+            dilation=(self.dilation, self.dilation),
+            batch_size=self.batch_size,
+            input_height=input_height,
+            input_width=input_width,
+            conv_config=self.conv_config,
+            compute_config=self.compute_config,
+            conv_op_cache=self.cache,
+            debug=False,
+            groups=self.groups,
+            memory_config=None,
+            return_weights_and_bias=True,
+            return_output_dim=True,
+        )
 
-    return (x, out_height, out_width)
+        if self.is_detect_cv2:
+            x = ttnn.sharded_to_interleaved(x, ttnn.L1_MEMORY_CONFIG)
+            return x, out_height, out_width
+
+        x = ttnn.silu(x)
+        return x, out_height, out_width
 
 
-def Bottleneck(
-    device,
-    x,
-    parameters,
-    path,
-    in_h,
-    in_w,
-    shortcut=True,
-    g=1,
-    k=(3, 3),
-    e=0.5,
-    act_block_h=True,
-    change_shard=None,
-    deallocate_activation=False,
-    output_layout=ttnn.TILE_LAYOUT,
-    tilize=False,
-):
-    cv1, out_h, out_w = conv(
+class Bottleneck:
+    def __init__(
+        self,
         device,
-        x,
         parameters,
-        f"{path}.cv1",
-        in_h,
-        in_w,
-        k[0][0],
-        1,
-        act_block_h=act_block_h,
-        change_shard=change_shard,
-        deallocate_activation=deallocate_activation,
-        output_layout=output_layout,
-    )
-
-    cv2, out_h, out_w = conv(
-        device,
-        cv1,
-        parameters,
-        f"{path}.cv2",
-        in_h,
-        in_w,
-        k[1][1],
-        1,
-        g=g,
-        act_block_h=act_block_h,
-        deallocate_activation=deallocate_activation,
-    )
-
-    ttnn.deallocate(cv1)
-
-    if tilize:
-        x = ttnn.to_layout(x, ttnn.TILE_LAYOUT, device=device, memory_config=ttnn.L1_MEMORY_CONFIG)
-
-    add = shortcut
-
-    return ttnn.add(x, cv2, memory_config=ttnn.L1_MEMORY_CONFIG) if add else cv2
-
-
-def c2f(
-    device,
-    x,
-    parameters,
-    path,
-    in_h,
-    in_w,
-    n=1,
-    shortcut=False,
-    g=1,
-    e=0.5,
-    act_block_h=False,
-    bfloat8=True,
-    block_shard=False,
-    change_shard=None,
-    deallocate_activation=False,
-    output_layout=ttnn.ROW_MAJOR_LAYOUT,
-    temp=False,
-):
-    cv1, out_h, out_w = conv(
-        device,
-        x,
-        parameters,
-        f"{path}.cv1",
-        in_h,
-        in_w,
-        1,
-        1,
-        bfloat8=bfloat8,
-        change_shard=change_shard,
-        deallocate_activation=deallocate_activation,
-        output_layout=output_layout,
-    )
-
-    cv1 = ttnn.to_layout(cv1, ttnn.ROW_MAJOR_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
-    y = list(ttnn.split(cv1, 2, 3, memory_config=ttnn.L1_MEMORY_CONFIG))
-    ttnn.deallocate(cv1)
-
-    to_tile = True
-
-    for i in range(n):
-        z = Bottleneck(
+        path,
+        shortcut,
+        change_shard,
+        input_params,
+        act_block_h=False,
+        deallocate_activation=False,
+        output_layout=ttnn.TILE_LAYOUT,
+        tilize=False,
+    ):
+        self.device = device
+        self.path = path
+        self.tilize = tilize
+        self.shortcut = shortcut
+        self.cv1 = Conv(
             device,
-            y[-1],
             parameters,
-            f"{path}.m.{i}",
-            out_h,
-            out_w,
-            shortcut,
-            g,
-            k=((3, 3), (3, 3)),
-            e=1.0,
+            f"{self.path}.cv1",
+            input_params,
+            change_shard=change_shard,
+            deallocate_activation=deallocate_activation,
+            output_layout=output_layout,
+        )
+        self.cv2 = Conv(
+            device,
+            parameters,
+            f"{self.path}.cv2",
+            input_params,
             act_block_h=act_block_h,
             change_shard=change_shard,
             deallocate_activation=deallocate_activation,
-            tilize=to_tile,
         )
-        y.append(z)
-        to_tile = False
 
-    y[0] = ttnn.to_layout(y[0], layout=ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
-    y[1] = ttnn.to_layout(y[1], layout=ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
+    def __call__(self, x):
+        cv1, out_h, out_w = self.cv1(x)
+        cv2, out_h, out_w = self.cv2(cv1)  # pass cv1
+        ttnn.deallocate(cv1)
 
-    if not shortcut:
-        for i in range(2, len(y)):
-            y[i] = ttnn.sharded_to_interleaved(y[i], ttnn.L1_MEMORY_CONFIG)
+        if self.tilize:
+            x = ttnn.to_layout(x, ttnn.TILE_LAYOUT, device=self.device, memory_config=ttnn.L1_MEMORY_CONFIG)
 
-    x = ttnn.concat(y, 3)  # memory_config=ttnn.L1_MEMORY_CONFIG OOM issue
+        return ttnn.add(x, cv2, memory_config=ttnn.L1_MEMORY_CONFIG) if self.shortcut else cv2
 
-    for i in range(len(y)):
-        ttnn.deallocate(y[i])
 
-    x, out_h, out_w = conv(
+class C2f:
+    def __init__(
+        self,
         device,
-        x,
         parameters,
-        f"{path}.cv2",
-        out_h,
-        out_w,
-        1,
-        bfloat8=bfloat8,
-        block_shard=block_shard,
-        change_shard=True,
-        deallocate_activation=deallocate_activation,
-    )
-    return x, out_h, out_w
+        path,
+        n=1,
+        shortcut=False,
+        change_shard=None,
+        input_params=None,
+        act_block_h=False,
+        bfloat8=True,
+        block_shard=False,
+        deallocate_activation=False,
+        output_layout=ttnn.ROW_MAJOR_LAYOUT,
+    ):
+        self.device = device
+        self.parameters = parameters
+        self.path = path
+        self.n = n
+        self.shortcut = shortcut
+        self.change_shard = change_shard
+        self.input_params = input_params
+        self.act_block_h = act_block_h
+        self.bfloat8 = bfloat8
+        self.block_shard = block_shard
+        self.deallocate_activation = deallocate_activation
+        self.output_layout = output_layout
+
+        self.cv1 = Conv(
+            device,
+            self.parameters,
+            f"{self.path}.cv1",
+            input_params=self.input_params[0],
+            bfloat8=self.bfloat8,
+            change_shard=self.change_shard,
+            deallocate_activation=self.deallocate_activation,
+            output_layout=self.output_layout,
+        )
+        self.cv2 = Conv(
+            self.device,
+            self.parameters,
+            f"{self.path}.cv2",
+            input_params=self.input_params[1],
+            bfloat8=self.bfloat8,
+            block_shard=self.block_shard,
+            change_shard=True,
+            deallocate_activation=self.deallocate_activation,
+        )
+        self.bottleneck_modules = []
+        for i in range(self.n):
+            if i == 0:
+                self.tilize = True
+            else:
+                self.tilize = False
+            self.bottleneck_modules.append(
+                Bottleneck(
+                    self.device,
+                    self.parameters,
+                    f"{self.path}.m.{i}",
+                    self.shortcut,
+                    self.change_shard,
+                    input_params=self.input_params[2],
+                    act_block_h=self.act_block_h,
+                    deallocate_activation=self.deallocate_activation,
+                    tilize=self.tilize,
+                )
+            )
+
+    def __call__(self, x):
+        cv1, out_h, out_w = self.cv1(x)
+        cv1 = ttnn.to_layout(cv1, ttnn.ROW_MAJOR_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
+
+        # y = list(
+        #     ttnn.split(cv1, 2, 3, memory_config=ttnn.L1_MEMORY_CONFIG)
+        # )  # use this for 320 resolution, but ttnn.split is not supprted by trace
+
+        # split is not supported by trace, hence using this this
+        y = []
+        y.append(cv1[:, :, :, : cv1.shape[-1] // 2])
+        y.append(cv1[:, :, :, cv1.shape[-1] // 2 :])
+
+        ttnn.deallocate(cv1)
+
+        to_tile = True
+        for i in range(self.n):
+            z = self.bottleneck_modules[i](y[-1])
+
+            y.append(z)
+            to_tile = False
+
+        y[0] = ttnn.to_layout(y[0], layout=ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
+        y[1] = ttnn.to_layout(y[1], layout=ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
+
+        if not self.shortcut:
+            for i in range(2, len(y)):
+                y[i] = ttnn.sharded_to_interleaved(y[i], ttnn.L1_MEMORY_CONFIG)
+
+        x = ttnn.concat(y, 3)
+
+        for i in range(len(y)):
+            ttnn.deallocate(y[i])
+
+        x, out_h, out_w = self.cv2(x)
+        return x, out_h, out_w
 
 
-def SPPF(device, x, parameters, path, in_h, in_w, k=5):
-    cv1, out_h, out_w = conv(device, x, parameters, f"{path}.cv1", in_h, in_w, 1, 1, change_shard=True)
+class SPPF:
+    def __init__(self, device, parameters, path, input_params, batch_size):
+        self.device = device
+        self.parameters = parameters
+        self.path = path
+        self.input_params = input_params
+        self.batch_size = batch_size
+        self.cv1 = Conv(device, parameters, f"{path}.cv1", input_params=input_params[0], change_shard=True)
+        self.cv2 = Conv(device, parameters, f"{path}.cv2", input_params=input_params[1], change_shard=True)
 
-    p = k // 2
+    def __call__(self, x):
+        cv1, out_h, out_w = self.cv1(x)
+        p = 5 // 2
+        cv1 = ttnn.to_layout(cv1, ttnn.ROW_MAJOR_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
+        y = [cv1]
+        for i in range(3):
+            output = ttnn.max_pool2d(
+                input_tensor=y[-1],
+                batch_size=self.batch_size,
+                input_h=out_h,
+                input_w=out_w,
+                channels=y[-1].shape[-1],
+                kernel_size=[5, 5],
+                stride=[1, 1],
+                padding=[p, p],
+                dilation=[1, 1],
+                memory_config=ttnn.L1_MEMORY_CONFIG,
+            )
+            y.append(output)
 
-    cv1 = ttnn.to_layout(cv1, ttnn.ROW_MAJOR_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
+        x = ttnn.concat(y, 3, memory_config=ttnn.L1_MEMORY_CONFIG)
+        for i in range(len(y)):
+            ttnn.deallocate(y[i])
 
-    y = [cv1]
-    for i in range(3):
-        output = ttnn.max_pool2d(
-            input_tensor=y[-1],
-            batch_size=y[-1].shape[0],
-            input_h=out_h,
-            input_w=out_w,
-            channels=y[-1].shape[-1],
-            kernel_size=[5, 5],
-            stride=[1, 1],
-            padding=[p, p],
-            dilation=[1, 1],
+        x, out_h, out_w = self.cv2(x)
+        return x, out_h, out_w
+
+
+class Detect_cv2:
+    def __init__(self, device, parameters, path, input_params):
+        self.device = device
+        self.parameters = parameters
+        self.path = path
+        self.input_params = input_params
+        self.conv0 = Conv(device, parameters, f"{path}.0", input_params=input_params[0], bfloat8=True)
+        self.conv1 = Conv(device, parameters, f"{path}.1", input_params=input_params[1], bfloat8=True)
+        self.conv2 = Conv(
+            device,
+            parameters,
+            path,
+            input_params=input_params[2],
+            bfloat8=True,
+            is_fused=False,
+            change_shard=True,
+            is_detect_cv2=True,
+        )
+
+    def __call__(self, x, reg_max=16):
+        x, out_h, out_w = self.conv0(x)
+        x, out_h, out_w = self.conv1(x)
+        x, out_h, out_w = self.conv2(x)
+        return x, out_h, out_w
+
+
+class DFL:
+    def __init__(self, device, parameters, path, input_params):
+        self.device = device
+        self.parameters = parameters
+        self.path = path
+        self.input_params = input_params
+        self.conv = Conv(
+            device, parameters, path, input_params, bfloat8=True, is_fused=False, is_dfl=True, change_shard=True
+        )
+
+    def __call__(self, x, c1=16):
+        b, _, a = x.shape
+
+        x = ttnn.reshape(x, (b, 4, c1, a), memory_config=ttnn.L1_MEMORY_CONFIG)
+
+        x = ttnn.permute(x, (0, 2, 1, 3), memory_config=ttnn.L1_MEMORY_CONFIG)
+
+        x = ttnn.to_layout(x, ttnn.TILE_LAYOUT)
+        x = ttnn.softmax(x, dim=1)
+
+        x = ttnn.permute(x, (0, 2, 3, 1), memory_config=ttnn.L1_MEMORY_CONFIG)
+        x, _, _ = self.conv(x)
+        x = ttnn.reshape(x, (x.shape[0], 4, -1))
+        return x
+
+
+class Detect:
+    def __init__(self, device, parameters, path, input_params, nc=80, ch=(320, 640, 640)):
+        self.device = device
+        self.parameters = parameters
+        self.path = path
+        self.input_params = input_params
+        self.nc = nc
+        self.ch = ch
+        self.detect_cv2_modules = []
+        self.detect_cv3_modules = []
+
+        nl = len(self.ch)
+        for i in range(nl):
+            cv2_params = input_params["cv2_params"][i]["input_params"]
+            cv3_params = input_params["cv3_params"][i]["input_params"]
+            self.detect_cv2_modules.append(Detect_cv2(device, parameters, f"{path}.cv2.{i}", input_params=cv2_params))
+            self.detect_cv3_modules.append(Detect_cv2(device, parameters, f"{path}.cv3.{i}", input_params=cv3_params))
+
+        self.dfl_module = DFL(
+            device, parameters, f"{path}.dfl", input_params=input_params["dfl_params"]["input_params"]
+        )
+
+    def __call__(self, x, nc=80, ch=()):
+        nc = self.nc
+        ch = self.ch
+        nl = len(ch)
+        reg_max = 16
+        no = nc + reg_max * 4
+
+        for i in range(nl):
+            a, _, _ = self.detect_cv2_modules[i](x[i])
+            b, _, _ = self.detect_cv3_modules[i](x[i])
+            x[i] = ttnn.concat((a, b), dim=3)
+
+        shape = x[0].shape
+        anchors, strides = self.parameters["anchors"], self.parameters["strides"]
+
+        xi = []
+        for i in x:
+            i = ttnn.reshape(i, (shape[0], -1, no), memory_config=ttnn.L1_MEMORY_CONFIG)
+            xi.append(i)
+
+        x_cat = ttnn.concat(xi, 1, memory_config=ttnn.L1_MEMORY_CONFIG)
+        x_cat = ttnn.permute(x_cat, (0, 2, 1), memory_config=ttnn.L1_MEMORY_CONFIG)
+
+        box = ttnn.slice(x_cat, [0, 0, 0], [1, 64, x_cat.shape[2]], memory_config=ttnn.L1_MEMORY_CONFIG)
+        cls = ttnn.slice(x_cat, [0, 64, 0], [1, 144, x_cat.shape[2]], memory_config=ttnn.L1_MEMORY_CONFIG)
+        dfl = self.dfl_module(box)
+        dbox = ttnn_decode_bboxes(self.device, dfl, anchors)
+        dbox = dbox * strides
+
+        return [ttnn.concat((dbox, ttnn.sigmoid(cls)), dim=1), x]
+
+
+class DetectionModel:
+    def __init__(self, device, parameters, res=(640, 640), batch_size=1):
+        self.device = device
+        self.parameters = parameters
+        self.res = res
+        self.reg_max = 16
+        self.batch_size = batch_size
+        sppf_configs = {"input_params": ((1, 1, 0, 320, 640), (1, 1, 0, 640, 1280))}
+        c2f_configs = {
+            "model.2": {"input_params": ((1, 1, 0, 160, 160), (1, 1, 0, 160, 400), (3, 1, 1, 80, 80))},
+            "model.4": {"input_params": ((1, 1, 0, 320, 320), (1, 1, 0, 320, 1280), (3, 1, 1, 160, 160))},
+            "model.6": {"input_params": ((1, 1, 0, 640, 640), (1, 1, 0, 640, 2560), (3, 1, 1, 320, 320))},
+            "model.8": {"input_params": ((1, 1, 0, 640, 640), (1, 1, 0, 640, 1600), (3, 1, 1, 320, 320))},
+            "model.12": {"input_params": ((1, 1, 0, 640, 1280), (1, 1, 0, 640, 1600), (3, 1, 1, 320, 320))},
+            "model.15": {"input_params": ((1, 1, 0, 320, 960), (1, 1, 0, 320, 800), (3, 1, 1, 160, 160))},
+            "model.18": {"input_params": ((1, 1, 0, 640, 960), (1, 1, 0, 640, 1600), (3, 1, 1, 320, 320))},
+            "model.21": {"input_params": ((1, 1, 0, 640, 1280), (1, 1, 0, 640, 1600), (3, 1, 1, 320, 320))},
+        }
+        detect_config = {
+            "cv2_params": [
+                {
+                    "input_params": [
+                        (3, 1, 1, 80, 320),
+                        (3, 1, 1, 80, 80),
+                        (1, 1, 0, 64, 80),
+                    ]
+                },
+                {
+                    "input_params": [
+                        (3, 1, 1, 80, 640),
+                        (3, 1, 1, 80, 80),
+                        (1, 1, 0, 64, 80),
+                    ]
+                },
+                {
+                    "input_params": [
+                        (3, 1, 1, 80, 640),
+                        (3, 1, 1, 80, 80),
+                        (1, 1, 0, 64, 80),
+                    ]
+                },
+            ],
+            "cv3_params": [
+                {
+                    "input_params": [
+                        (3, 1, 1, 320, 320),
+                        (3, 1, 1, 320, 320),
+                        (1, 1, 0, 80, 320),
+                    ]
+                },
+                {
+                    "input_params": [
+                        (3, 1, 1, 320, 640),
+                        (3, 1, 1, 320, 320),
+                        (1, 1, 0, 80, 320),
+                    ]
+                },
+                {
+                    "input_params": [
+                        (3, 1, 1, 320, 640),
+                        (3, 1, 1, 320, 320),
+                        (1, 1, 0, 80, 320),
+                    ]
+                },
+            ],
+            "dfl_params": {
+                "input_params": (1, 1, 0, 1, 16),
+            },
+        }
+
+        self.conv_0 = Conv(
+            device,
+            parameters,
+            "model.0",
+            input_params=[3, 2, 1, 80, 3],
+            change_shard=False,
+            deallocate_activation=True,
+            act_block_h=True,
+            is_dfl=True,
+        )
+        self.conv_1 = Conv(
+            device, parameters, "model.1", input_params=[3, 2, 1, 160, 80], act_block_h=True, deallocate_activation=True
+        )
+        self.c2f_2 = C2f(
+            device,
+            parameters,
+            "model.2",
+            n=3,
+            shortcut=True,
+            change_shard=False,
+            input_params=c2f_configs["model.2"]["input_params"],
+        )
+        self.conv_3 = Conv(
+            device,
+            parameters,
+            "model.3",
+            input_params=[3, 2, 1, 320, 160],
+            change_shard=True,
+            deallocate_activation=False,
+        )
+        self.c2f_4 = C2f(
+            device,
+            parameters,
+            "model.4",
+            n=6,
+            shortcut=True,
+            change_shard=False,
+            input_params=c2f_configs["model.4"]["input_params"],
+        )
+        self.conv_5 = Conv(
+            device, parameters, "model.5", input_params=[3, 2, 1, 640, 320], change_shard=False, block_shard=False
+        )
+        self.c2f_6 = C2f(
+            device,
+            parameters,
+            "model.6",
+            n=6,
+            shortcut=True,
+            block_shard=False,
+            change_shard=False,
+            input_params=c2f_configs["model.6"]["input_params"],
+        )
+        self.conv_7 = Conv(
+            device, parameters, "model.7", input_params=[3, 2, 1, 640, 640], block_shard=False, change_shard=True
+        )
+        self.c2f_8 = C2f(
+            device,
+            parameters,
+            "model.8",
+            n=3,
+            shortcut=True,
+            change_shard=True,
+            input_params=c2f_configs["model.8"]["input_params"],
+        )
+        self.sppf_9 = SPPF(
+            device, parameters, "model.9", input_params=sppf_configs["input_params"], batch_size=self.batch_size
+        )
+        self.c2f_12 = C2f(
+            device,
+            parameters,
+            "model.12",
+            n=3,
+            shortcut=False,
+            bfloat8=True,
+            input_params=c2f_configs["model.12"]["input_params"],
+        )
+        self.c2f_15 = C2f(
+            device, parameters, "model.15", n=3, shortcut=False, input_params=c2f_configs["model.15"]["input_params"]
+        )
+        self.conv_16 = Conv(device, parameters, "model.16", input_params=[3, 2, 1, 320, 320])
+        self.c2f_18 = C2f(
+            device, parameters, "model.18", n=3, shortcut=False, input_params=c2f_configs["model.18"]["input_params"]
+        )
+        self.conv_19 = Conv(device, parameters, "model.19", input_params=[3, 2, 1, 640, 640], block_shard=True)
+        self.c2f_21 = C2f(
+            device,
+            parameters,
+            "model.21",
+            n=3,
+            shortcut=False,
+            input_params=c2f_configs["model.21"]["input_params"],
+            change_shard=True,
+        )
+        self.detect_22 = Detect(device, parameters, "model.22", detect_config)
+
+    def __call__(self, x):
+        conv_0, out_h, out_w = self.conv_0(x)
+        # ttnn.deallocate(x)
+        conv_1, out_h, out_w = self.conv_1(conv_0)
+        ttnn.deallocate(conv_0)
+        c2f_2, out_h, out_w = self.c2f_2(conv_1)
+        ttnn.deallocate(conv_1)
+        conv_3, out_h, out_w = self.conv_3(c2f_2)
+        ttnn.deallocate(c2f_2)
+        c2f_4, out_h, out_w = self.c2f_4(conv_3)
+        ttnn.deallocate(conv_3)
+        c2f_4 = ttnn.sharded_to_interleaved(c2f_4, ttnn.L1_MEMORY_CONFIG)
+        c2f_4 = ttnn.reallocate(c2f_4, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        conv_5, out_h, out_w = self.conv_5(c2f_4)
+        c2f_6, out_h, out_w = self.c2f_6(conv_5)
+        ttnn.deallocate(conv_5)
+        c2f_6 = ttnn.reallocate(c2f_6, memory_config=ttnn.L1_MEMORY_CONFIG)
+        conv_7, out_h, out_w = self.conv_7(c2f_6)
+        conv_7 = ttnn.sharded_to_interleaved(conv_7, ttnn.L1_MEMORY_CONFIG)
+        c2f_8, out_h, out_w = self.c2f_8(conv_7)
+        ttnn.deallocate(conv_7)
+        nine, out_h, out_w = self.sppf_9(c2f_8)
+        ttnn.deallocate(c2f_8)
+        SPPF_9 = ttnn.to_layout(nine, ttnn.ROW_MAJOR_LAYOUT)
+        SPPF_9 = ttnn.reshape(SPPF_9, (self.batch_size, out_h, out_w, SPPF_9.shape[-1]))
+        SPPF_9 = ttnn.upsample(SPPF_9, scale_factor=(2, 2))
+        x = ttnn.reshape(
+            SPPF_9,
+            (1, 1, (self.batch_size) * SPPF_9.shape[1] * SPPF_9.shape[2], SPPF_9.shape[-1]),
             memory_config=ttnn.L1_MEMORY_CONFIG,
         )
-        y.append(output)
-
-    x = ttnn.concat(y, 3, memory_config=ttnn.L1_MEMORY_CONFIG)
-
-    for i in range(len(y)):
-        ttnn.deallocate(y[i])
-
-    x, out_h, out_w = conv(device, x, parameters, f"{path}.cv2", out_h, out_w, 1, 1, change_shard=True)
-
-    return x, out_h, out_w
-
-
-def Detect_cv2(device, x, parameters, path, inp_h, inp_w, k, reg_max, bfloat8=True):
-    x, out_h, out_w = conv(device, x, parameters, f"{path}.0", inp_h, inp_w, k, bfloat8=bfloat8)
-
-    x, out_h, out_w = conv(device, x, parameters, f"{path}.1", out_h, out_w, k, bfloat8=bfloat8)
-
-    x, out_h, out_w = conv(
-        device,
-        x,
-        parameters,
-        path,
-        out_h,
-        out_w,
-        k=1,
-        s=1,
-        p=0,
-        g=1,
-        d=1,
-        bfloat8=True,
-        is_fused=False,
-        change_shard=True,
-        is_detect_cv2=True,
-    )
-    return x, out_h, out_w
-
-
-def DFL(device, x, parameters, path, c1=16):
-    b, _, a = x.shape
-
-    x = ttnn.reshape(x, (b, 4, c1, a))
-
-    x = ttnn.softmax(x, dim=2)
-
-    x = ttnn.permute(x, (0, 1, 3, 2))
-
-    x = conv(
-        device,
-        x,
-        parameters,
-        path,
-        x.shape[1],
-        x.shape[2],
-        k=1,
-        s=1,
-        p=0,
-        g=1,
-        d=1,
-        bfloat8=True,
-        is_fused=False,
-        is_dfl=True,
-        change_shard=True,
-    )
-
-    return x
-
-
-def Detect(device, x, parameters, path, nc=80, ch=(), bfloat8=True):
-    nc = nc
-    nl = len(ch)
-    reg_max = 16
-    no = nc + reg_max * 4
-
-    for i in range(nl):
-        inp_h = inp_w = int(math.sqrt(x[i].shape[2]))
-        a = Detect_cv2(
-            device,
-            x[i],
-            parameters,
-            path=f"{path}.cv2.{i}",
-            inp_h=inp_h,
-            inp_w=inp_w,
-            k=3,
-            reg_max=4 * reg_max,
-        )[0]
-        b = Detect_cv2(
-            device,
-            x[i],
-            parameters,
-            path=f"{path}.cv3.{i}",
-            inp_h=inp_h,
-            inp_w=inp_w,
-            k=3,
-            reg_max=nc,
-            bfloat8=bfloat8,
-        )[0]
-        x[i] = ttnn.concat((a, b), dim=3)
-
-    shape = x[0].shape
-
-    anchors, strides = parameters["anchors"], parameters["strides"]
-
-    xi = []
-    for i in x:
-        i = ttnn.reshape(i, (shape[0], -1, no))
-        xi.append(i)
-
-    x_cat = ttnn.concat(xi, 1)
-
-    x_cat = ttnn.permute(x_cat, (0, 2, 1))
-
-    box = ttnn.slice(x_cat, [0, 0, 0], [1, 64, x_cat.shape[2]])
-    cls = ttnn.slice(x_cat, [0, 64, 0], [1, 144, x_cat.shape[2]])
-
-    dfl = DFL(device, box, parameters, f"{path}.dfl")
-
-    dbox = ttnn_decode_bboxes(device, dfl, anchors)
-    dbox = dbox * strides
-
-    return [ttnn.concat((dbox, ttnn.sigmoid(cls)), dim=1), x]
-
-
-def DetectionModel(device, x, parameters, res):
-    conv_0, out_h, out_w = conv(
-        device, x, parameters, "model.0", res[0], res[1], 3, 2, 1, change_shard=True, deallocate_activation=True
-    )
-    ttnn.deallocate(x)
-
-    conv_1, out_h, out_w = conv(
-        device, conv_0, parameters, "model.1", out_h, out_w, 3, 2, 1, act_block_h=True, deallocate_activation=True
-    )
-    ttnn.deallocate(conv_0)
-
-    c2f_2, out_h, out_w = c2f(
-        device, conv_1, parameters, "model.2", out_h, out_w, n=3, shortcut=True, change_shard=False
-    )
-    ttnn.deallocate(conv_1)
-
-    conv_3, out_h, out_w = conv(
-        device, c2f_2, parameters, "model.3", out_h, out_w, 3, 2, 1, change_shard=True, deallocate_activation=False
-    )
-    ttnn.deallocate(c2f_2)
-
-    c2f_4, out_h, out_w = c2f(
-        device, conv_3, parameters, "model.4", out_h, out_w, n=6, shortcut=True, change_shard=False
-    )
-    ttnn.deallocate(conv_3)
-
-    c2f_4 = ttnn.sharded_to_interleaved(c2f_4, ttnn.L1_MEMORY_CONFIG)
-
-    c2f_4 = ttnn.reallocate(c2f_4, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-
-    conv_5, out_h, out_w = conv(
-        device, c2f_4, parameters, "model.5", out_h, out_w, 3, 2, 1, change_shard=False, block_shard=False
-    )
-
-    c2f_6, out_h, out_w = c2f(
-        device, conv_5, parameters, "model.6", out_h, out_w, n=6, shortcut=True, block_shard=False, change_shard=False
-    )
-    ttnn.deallocate(conv_5)
-
-    c2f_6 = ttnn.reallocate(c2f_6, memory_config=ttnn.L1_MEMORY_CONFIG)
-
-    conv_7, out_h, out_w = conv(
-        device,
-        c2f_6,
-        parameters,
-        "model.7",
-        out_h,
-        out_w,
-        3,
-        2,
-        1,
-        block_shard=False,
-        change_shard=True,
-    )
-
-    conv_7 = ttnn.sharded_to_interleaved(conv_7, ttnn.L1_MEMORY_CONFIG)
-
-    c2f_8, out_h, out_w = c2f(
-        device, conv_7, parameters, "model.8", out_h, out_w, n=3, shortcut=True, change_shard=True
-    )
-    ttnn.deallocate(conv_7)
-
-    nine, out_h, out_w = SPPF(device, c2f_8, parameters, "model.9", out_h, out_w)
-    ttnn.deallocate(c2f_8)
-
-    SPPF_9 = ttnn.to_layout(nine, ttnn.ROW_MAJOR_LAYOUT)
-
-    SPPF_9 = ttnn.reshape(SPPF_9, (1, out_h, out_w, SPPF_9.shape[-1]))
-    SPPF_9 = ttnn.upsample(SPPF_9, scale_factor=(2, 2))
-
-    x = ttnn.reshape(SPPF_9, (1, 1, SPPF_9.shape[1] * SPPF_9.shape[2], SPPF_9.shape[-1]))
-    x = ttnn.to_layout(x, ttnn.TILE_LAYOUT)
-
-    x = ttnn.concat([x, c2f_6], dim=3)
-    ttnn.deallocate(c2f_6)
-
-    c2f_12, out_h, out_w = c2f(
-        device, x, parameters, "model.12", SPPF_9.shape[1], SPPF_9.shape[2], n=3, shortcut=False, bfloat8=True
-    )
-    ttnn.deallocate(x)
-    ttnn.deallocate(SPPF_9)
-
-    c2f_12 = ttnn.sharded_to_interleaved(c2f_12, ttnn.L1_MEMORY_CONFIG)
-    c2f_12 = ttnn.to_layout(c2f_12, ttnn.ROW_MAJOR_LAYOUT)
-
-    twelve = ttnn.clone(c2f_12, dtype=ttnn.bfloat16, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-
-    c2f_12 = ttnn.reshape(c2f_12, (1, out_h, out_w, c2f_12.shape[-1]))
-    c2f_12 = ttnn.upsample(c2f_12, scale_factor=(2, 2))
-
-    x = ttnn.reshape(c2f_12, (1, 1, c2f_12.shape[1] * c2f_12.shape[2], c2f_12.shape[-1]))
-    x = ttnn.to_layout(x, ttnn.TILE_LAYOUT)
-    x = ttnn.concat([x, c2f_4], dim=3)
-    ttnn.deallocate(c2f_4)
-    ttnn.deallocate(c2f_12)
-
-    c2f_15, out_h, out_w = c2f(device, x, parameters, "model.15", c2f_12.shape[1], c2f_12.shape[2], n=3, shortcut=False)
-    ttnn.deallocate(x)
-
-    c2f_15 = ttnn.sharded_to_interleaved(c2f_15, ttnn.L1_MEMORY_CONFIG)
-    fifteen = ttnn.clone(c2f_15, dtype=ttnn.bfloat16, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-
-    conv_16, out_h, out_w = conv(device, c2f_15, parameters, "model.16", out_h, out_w, 3, 2, 1)
-    ttnn.deallocate(c2f_15)
-    conv_16 = ttnn.sharded_to_interleaved(conv_16, ttnn.L1_MEMORY_CONFIG)
-    conv_16 = ttnn.to_layout(conv_16, ttnn.ROW_MAJOR_LAYOUT)
-
-    x = ttnn.concat([conv_16, twelve], dim=3)
-    ttnn.deallocate(twelve)
-    ttnn.deallocate(conv_16)
-
-    c2f_18, out_h, out_w = c2f(device, x, parameters, "model.18", out_h, out_w, n=3, shortcut=False)
-    ttnn.deallocate(x)
-
-    c2f_18 = ttnn.sharded_to_interleaved(c2f_18, ttnn.L1_MEMORY_CONFIG)
-    eighteen = ttnn.clone(c2f_18, dtype=ttnn.bfloat16, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-
-    conv_19, out_h, out_w = conv(device, c2f_18, parameters, "model.19", out_h, out_w, 3, 2, 1, block_shard=True)
-    ttnn.deallocate(c2f_18)
-    conv_19 = ttnn.sharded_to_interleaved(conv_19, ttnn.L1_MEMORY_CONFIG)
-
-    x = ttnn.concat([conv_19, nine], dim=3)
-    ttnn.deallocate(nine)
-    ttnn.deallocate(conv_19)
-
-    c2f_21, out_h, out_w = c2f(device, x, parameters, "model.21", out_h, out_w, n=3, shortcut=False)
-    ttnn.deallocate(x)
-
-    x = [fifteen, eighteen, c2f_21]
-
-    x = Detect(device, x, parameters, "model.22", nc=80, ch=(320, 640, 640))
-
-    return x
-
-
-def YOLOv8x(device, x, parameters, res=(320, 320)):
-    return DetectionModel(device, x, parameters, res)
+        x = ttnn.to_layout(x, ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
+        x = ttnn.concat([x, c2f_6], dim=3)
+        ttnn.deallocate(c2f_6)
+        c2f_12, out_h, out_w = self.c2f_12(x)
+        ttnn.deallocate(x)
+        ttnn.deallocate(SPPF_9)
+        c2f_12 = ttnn.sharded_to_interleaved(c2f_12, ttnn.L1_MEMORY_CONFIG)
+        c2f_12 = ttnn.to_layout(c2f_12, ttnn.ROW_MAJOR_LAYOUT)
+        twelve = ttnn.clone(c2f_12, dtype=ttnn.bfloat16, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        c2f_12 = ttnn.reshape(
+            c2f_12, (self.batch_size, out_h, out_w, c2f_12.shape[-1]), memory_config=ttnn.L1_MEMORY_CONFIG
+        )
+        c2f_12 = ttnn.upsample(c2f_12, scale_factor=(2, 2))
+        x = ttnn.reshape(c2f_12, (1, 1, (self.batch_size) * c2f_12.shape[1] * c2f_12.shape[2], c2f_12.shape[-1]))
+        x = ttnn.to_layout(x, ttnn.TILE_LAYOUT)
+        x = ttnn.concat([x, c2f_4], dim=3)
+        ttnn.deallocate(c2f_4)
+        ttnn.deallocate(c2f_12)
+        c2f_15, out_h, out_w = self.c2f_15(x)
+        ttnn.deallocate(x)
+        c2f_15 = ttnn.sharded_to_interleaved(c2f_15, ttnn.L1_MEMORY_CONFIG)
+        fifteen = ttnn.clone(c2f_15, dtype=ttnn.bfloat16, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        conv_16, out_h, out_w = self.conv_16(c2f_15)
+        ttnn.deallocate(c2f_15)
+        conv_16 = ttnn.sharded_to_interleaved(conv_16, ttnn.L1_MEMORY_CONFIG)
+        conv_16 = ttnn.to_layout(conv_16, ttnn.ROW_MAJOR_LAYOUT)
+        x = ttnn.concat([conv_16, twelve], dim=3)
+        ttnn.deallocate(twelve)
+        ttnn.deallocate(conv_16)
+        c2f_18, out_h, out_w = self.c2f_18(x)
+        ttnn.deallocate(x)
+        c2f_18 = ttnn.sharded_to_interleaved(c2f_18, ttnn.L1_MEMORY_CONFIG)
+        eighteen = ttnn.clone(c2f_18, dtype=ttnn.bfloat16, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        conv_19, out_h, out_w = self.conv_19(c2f_18)
+        ttnn.deallocate(c2f_18)
+        conv_19 = ttnn.sharded_to_interleaved(conv_19, ttnn.L1_MEMORY_CONFIG)
+        x = ttnn.concat([conv_19, nine], dim=3)
+        ttnn.deallocate(nine)
+        ttnn.deallocate(conv_19)
+        c2f_21, out_h, out_w = self.c2f_21(x)
+        ttnn.deallocate(x)
+        x = [fifteen, eighteen, c2f_21]
+        x = self.detect_22(x, nc=80, ch=(320, 640, 640))
+        return x
+
+
+class YOLOv8xModel:
+    def __init__(self, device, parameters, res=(640, 640), batch_size=1):
+        self.device = device
+        self.parameters = parameters
+        self.res = res
+        self.detection_model = DetectionModel(device, parameters, res, batch_size)
+
+    def __call__(self, x):
+        return self.detection_model(x)
