+++ tests/scripts/single_card/run_single_card_demo_tests.sh
@@ -61,6 +61,10 @@ run_common_func_tests() {
   #RoBERTa
   pytest --disable-warnings models/demos/roberta/demo/demo.py --timeout 600; fail+=$?
 
+  #yolov8x
+  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest --disable-warnings models/experimental/functional_yolov8x/demo/demo.py --timeout 600; fail+=$?
+
+
   return $fail
 }
